refactor(web): inline annotate drawer trigger components
